230824/woorek load_BoxOffice에 박스오피스 지역코드 DB insert function추가
